changes in Obstacle.java and Pos.java (double to int)

+++ src/fourmisses/Pos.java
@@ -2,11 +2,11 @@ package fourmisses;
 
 abstract public class Pos {
 
-	double x;
-	double y;
+	int x;
+	int y;
 	
 	/////CONSTRUCTOR/////
-	public Pos(double x, double y) {
+	public Pos(int x, int y) {
 		super();
 		this.x = x;
 		this.y = y;
@@ -14,16 +14,16 @@ abstract public class Pos {
 	
 	
 	/////METHODS/////
-	public double getX() {
+	public int getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
-	public double getY() {
+	public int getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
